// src/types/mysql.d.ts
export interface FilesInfo {
    // 源表名称: files_info - 主键: `file_id` DESC
    /** 文件名 必填   */
    file_name: string;
    /** 文件id 可选 - 自增长  */
    file_id?: number;
    /** 文件路径 必填   */
    file_path: string;
    /** 文件后缀 必填   */
    file_ext: string;
    /** 文件上传时间 可选  - 默认 */
    upload_time?: string;
    /** 文件的类型（MIME类型） 必填   */
    file_type: string;
    /** 文件的大小 必填   */
    file_size: number;
    /** 文件全名 必填   */
    file_fullname: string;
    /** 用户id 可选  - 默认 */
    user_id?: string;
    /** 文件的哈希值 必填   */
    hash: string;
    /** 文件的状态 必填   */
    status: 'active' | 'inactive' | 'deleted';
    /** 文件的描述 可选   */
    description?: string;
}